+++ example/lib/main.dart
@@ -1,54 +1,136 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:flutter_flexible_toast/flutter_flexible_toast.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(new MyApp());
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => _MyAppState();
+  _MyAppState createState() => new _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
-
   @override
-  void initState() {
+  initState() {
     super.initState();
-    initPlatformState();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    try {
-      platformVersion = await FlutterFlexibleToast.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
+  void showLongToast() {
+    FlutterFlexibleToast.showToast(
+      message: "Long Simple Toast",
+      toastLength: Toast.LENGTH_LONG,
+    );
+  }
 
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
+  void showColoredToast() {
+    FlutterFlexibleToast.showToast(
+        message: "Short red Success Toast",
+        toastLength: Toast.LENGTH_SHORT,
+        backgroundColor: Colors.red,
+        icon: ICON.SUCCESS,
+        fontSize: 16,
+        textColor: Colors.white);
+  }
 
-    setState(() {
-      _platformVersion = platformVersion;
-    });
+  void showShortToast() {
+    FlutterFlexibleToast.showToast(
+        message: "Short Info 1 Sec Toast",
+        toastLength: Toast.LENGTH_SHORT,
+        icon: ICON.INFO,
+        timeInSeconds: 1);
+  }
+
+  void showTopShortToast() {
+    FlutterFlexibleToast.showToast(
+        message: "Top Short Warning 1 Sec Toast",
+        toastLength: Toast.LENGTH_SHORT,
+        toastGravity: ToastGravity.TOP,
+        icon: ICON.WARNING,
+        timeInSeconds: 1);
+  }
+
+  void showCenterShortToast() {
+    FlutterFlexibleToast.showToast(
+        message: "Center Short Warning 1 Sec Toast",
+        toastLength: Toast.LENGTH_SHORT,
+        toastGravity: ToastGravity.CENTER,
+        icon: ICON.WARNING,
+        timeInSeconds: 1);
+  }
+
+  void showCenterShortLoadingToast() {
+    FlutterFlexibleToast.showToast(
+        message: "Short Loading 2 Sec Toast",
+        toastLength: Toast.LENGTH_LONG,
+        toastGravity: ToastGravity.BOTTOM,
+        icon: ICON.LOADING,
+        radius: 100,
+        elevation: 10,
+        textColor: Colors.white,
+        backgroundColor: Colors.black,
+        timeInSeconds: 2);
+  }
+
+  void cancelToast() {
+    FlutterFlexibleToast.cancel();
   }
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Plugin example app'),
+    return new MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: new Scaffold(
+        appBar: new AppBar(
+          title: new Text('Flutter Flexible Toast'),
         ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+        body: new Center(
+          child: new Column(
+            children: <Widget>[
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Long Simple Toast'),
+                    onPressed: showLongToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Short Info 1 Sec Toast'),
+                    onPressed: showShortToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Center Short Warning 1 Sec Toast'),
+                    onPressed: showCenterShortToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Top Short Warning 1 Sec Toast'),
+                    onPressed: showTopShortToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Short red Success Toast'),
+                    onPressed: showColoredToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                    child: new Text('Short Loading 2 Sec Toast'),
+                    onPressed: showCenterShortLoadingToast),
+              ),
+              new Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: new RaisedButton(
+                  child: new Text('Cancel Toasts'),
+                  onPressed: cancelToast,
+                ),
+              ),
+            ],
+          ),
         ),
       ),
     );